+++ src/Modules/Region/ForumTransactions.php
@@ -151,8 +151,10 @@ class ForumTransactions
             $moderators = $this->foodsaverGateway->getAdminsOrAmbassadors($moderationGroup);
         }
         if ($moderators) {
+            // send notification e-mail
+            $link = BASE_URL . $this->url($region['id'], false, $threadId);
             $data = [
-                'link' => BASE_URL . $this->url($region['id'], false, $threadId),
+                'link' => $link,
                 'thread' => $thread['title'],
                 'post' => $this->sanitizerService->markdownToHtml($rawPostBody),
                 'poster' => $posterName,
@@ -160,6 +162,22 @@ class ForumTransactions
             ];
 
             $this->sendNotificationMail($moderators, 'forum/activation', $data);
+
+            // create notification bell
+            $bellData = Bell::create(
+                'forum_not_activated_thread_title',
+                'forum_not_activated_thread',
+                'fas fa-comments',
+                ['href' => $link],
+                [
+                    'user' => $this->session->user('name'),
+                    'forum' => $region['name'],
+                    'title' => $thread['title'],
+                ],
+                BellType::createIdentifier(BellType::NOT_ACTIVATED_FORUM_THREAD, $threadId),
+                false,
+            );
+            $this->bellGateway->addBell(array_column($moderators, 'id'), $bellData);
         }
     }
 
@@ -278,5 +296,38 @@ class ForumTransactions
         );
 
         $this->bellGateway->addBell($notifiedUsers, $bell);
+    }
+
+    /**
+     * Activates a thread in a moderated forum. This function does nothing if the thread is already activated.
+     */
+    public function activateThread(int $threadId): void
+    {
+        $this->forumGateway->activateThread($threadId);
+        $this->removeInactiveThreadBell($threadId);
+    }
+
+    /**
+     * Deletes a thread. Removes the corresponding bell notifications, if the thread was not yet activated. This
+     * function does nothing if the thread does not exist.
+     */
+    public function deleteThread(int $threadId): void
+    {
+        $this->forumGateway->deleteThread($threadId);
+        $this->removeInactiveThreadBell($threadId);
+    }
+
+    /**
+     * Removes the bell that was created to notify moderators about a new thread. This function does nothing if
+     * the thread or the bell do not exist.
+     *
+     * @param int $threadId the thread for which the bell was created
+     */
+    private function removeInactiveThreadBell(int $threadId): void
+    {
+        $identifier = BellType::createIdentifier(BellType::NOT_ACTIVATED_FORUM_THREAD, $threadId);
+        if ($this->bellGateway->bellWithIdentifierExists($identifier)) {
+            $this->bellGateway->delBellsByIdentifier($identifier);
+        }
     }
 }